+++ inst/examples/smc_post_processing_functions_example.R
@@ -23,21 +23,13 @@ smc_test <- smc_mallows_new_users(
   alpha_max = 1e6
 )
 
-test_sample_rho <- smc_test$rho_samples[, , Time + 1]
-compute_posterior_intervals_rho(
-  output = test_sample_rho,
-  nmc = N, burnin = 0,
-  verbose = FALSE
-)
+compute_posterior_intervals(smc_test, parameter = "rho")
 
+test_sample_rho <- smc_test$rho_samples[, , Time + 1]
 compute_rho_consensus(
   output = test_sample_rho, nmc = N,
   burnin = 0, C = 1, type = "CP",
   verbose = FALSE
 )
 
-test_sample_alpha <- smc_test$alpha_samples[, Time + 1]
-compute_posterior_intervals_alpha(
-  output = test_sample_alpha,
-  nmc = N, burnin = 0, verbose = FALSE
-)
+compute_posterior_intervals(smc_test, parameter = "alpha")