+++ src/smc_correction_kernel_pseudo.cpp
@@ -22,6 +22,7 @@ using namespace arma;
 //'   \code{"ulam"}.
 //' @return list containing R_obs, the proposed 'corrected' augmented ranking that is compatible with the new observed ranking for a user, and
 //'         forward_auxiliary_ranking_probability, a numerical value for the probability of correcting the ranking to be compatible with R_obs.
+//' @keywords internal
 // [[Rcpp::export]]
 Rcpp::List correction_kernel_pseudo(
     const arma::vec current_ranking,  //R_curr