+++ src/smc_get_sample_probabilities.cpp
@@ -19,7 +19,8 @@ using namespace arma;
 //' @param n_items Integer is the number of items in the consensus ranking
 //' @return sample_prob_list A numeric sequence of sample probabilities for selecting a specific rank given the current
 //'         rho_item_rank
-//' @export
+//' @noRd
+//'
 // [[Rcpp::export]]
 arma::vec get_sample_probabilities(
   const arma::vec rho_item_rank,