+++ src/smc_metropolis_hastings_aug_ranking.cpp
@@ -20,7 +20,7 @@ using namespace arma;
 //'   \code{"ulam"}.
 //' @param pseudo Boolean specifying whether to use pseudo proposal or not.s
 //' @return R_curr or R_obs A ranking sequence vector representing proposed augmented ranking for next iteration of MCMC chain
-//' @export
+//' @noRd
 // [[Rcpp::export]]
 arma::vec metropolis_hastings_aug_ranking(
 	const double& alpha,