+++ tests/testthat/test-smc_mallows_complete_rankings.R
@@ -71,9 +71,7 @@ test_that("Output of smc_mallows_new_users_complete is OK", {
 # ###############################
 
 # posterior confidence intervals for rho
-rho_temp <- compute_posterior_intervals_rho(
-  output = test$rho_samples[, , Time + 1], nmc = N, burnin = 0
-)
+rho_temp <- compute_posterior_intervals(test, parameter = "rho")
 
 # MAP AND CP consensus ranking estimates
 rho_cp <- compute_rho_consensus(
@@ -99,8 +97,8 @@ alpha_samples_table <- data.frame(
   iteration = 1:N, value = test$alpha_samples[, Time + 1]
 )
 # posterior confidence intervals
-alpha_posterior_intervals <- compute_posterior_intervals_alpha(
-  output = test$alpha_samples[, Time + 1], nmc = N, burnin = 0
+alpha_posterior_intervals <- compute_posterior_intervals(
+  test, parameter = "alpha"
 )
 
 test_that("Output of compute_posterior_intervals_alpha is OK", {